reorganized code and implemented sign out button
