+++ app/src/main/java/com/example/appdev2_assignment2/ui/UserProfilePage.kt
@@ -13,7 +13,9 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
@@ -30,10 +32,13 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
 import com.example.appdev2_assignment2.AppUser
 import com.example.appdev2_assignment2.R
 import com.example.appdev2_assignment2.User
 import com.example.appdev2_assignment2.ViewModels.UserViewModel
+import com.example.appdev2_assignment2.auth.signOut
+import com.google.firebase.auth.FirebaseAuth
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -43,7 +48,9 @@ fun UserProfilePage(
     onNameChange: () -> Unit,
     onAgeChange: () -> Unit,
     onPasswordChange: () -> Unit,
-    onApplyChanges: () -> Unit
+    onApplyChanges: () -> Unit,
+    navController: NavController,
+    auth: FirebaseAuth
 )  {
 
     var isChangingName by remember { mutableStateOf(false) }
@@ -62,7 +69,8 @@ fun UserProfilePage(
     Column(
         modifier = Modifier
             .fillMaxSize()
-            .padding(16.dp),
+            .padding(16.dp)
+            .verticalScroll(rememberScrollState()),
         verticalArrangement = Arrangement.spacedBy(16.dp)
     ) {
         // Profile Picture
@@ -131,6 +139,17 @@ fun UserProfilePage(
             modifier = Modifier.fillMaxWidth()
         ) {
             Text("Change Age")
+        }
+
+        Button(
+            onClick = {
+                signOut(auth, navController)
+            },
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(8.dp)
+        ){
+            Text(text = "Sign Out")
         }
